move config to own package
