+++ bot/__init__.py
@@ -1,4 +1,3 @@
 from .TS3Bot import Bot
-from .config import Config
 
-__all__ = ['Config', 'Bot']
+__all__ = ['Bot']
